refactor: create separate components for pie chart based on screen size
